% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncronise_functions.R
\name{autofit}
\alias{autofit}
\title{Autofit function for calibration model}
\usage{
autofit(
  t = tekextract(),
  s = simextract(),
  attempts = 6,
  cycles = 5,
  target = 0.97
)
}
\arguments{
\item{t}{Output of the tekextract() function to feed into the indexer() function}

\item{s}{Output of the simextract() function to feed into the indexer() function}

\item{attempts}{The number of different peaks the matching algorithm should iterate through}

\item{cycles}{The number of cycles recorded, assumed to be 5}

\item{target}{The target R squared value to meet or exceed before returning}
}
\value{

}
\description{
This function iteratively controls the indexer parameters to optimise the calibration model
There are two loops, one which sets the peak number between 1 and the value of attempts,
and another which searches a small adjustment modifier from -10 to +10.
Stops when the calibration model fits to an R squared of >the target value.
Has three outputs to the global environment (using the tidytek and calmodel functions):
model
tek_sum
tek_ind
}
\examples{
\dontrun{
autofit()
}
}
