% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{calibrate}
\alias{calibrate}
\title{Pressure map calibration}
\usage{
calibrate(file = sampleextract(), modselect = model)
}
\arguments{
\item{file}{TekScan output to be calibrated. Defaults to the output of sampleextract()}

\item{modselect}{the model to be used for prediction. Defaults to the output of autofit() from the GlobalEnvironment}
}
\value{

}
\description{
This function uses the model to calculate the MPa value of each sensel in the pressure map.
Used as a helper function for other outputs, but can be called to produce results as a dataframe.
}
\examples{
\dontrun{
calibrate()
}
}
