% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncronise_functions.R
\name{calmodel}
\alias{calmodel}
\title{Calibration model generator}
\usage{
calmodel(tek = tek_sum, sim = sim_sum)
}
\arguments{
\item{tek}{The synchronised TekScan summary output. One output of the indexer() function}

\item{sim}{The output of simextract()}
}
\value{

}
\description{
Performs a polynomial regression to generate an equation converting the raw TekScan values into Newtons.
required to calibrate the output of the TekScan
}
\examples{
\dontrun{
calmodel(tek_sum, sim_sum)
}
}
