% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{featbarplot}
\alias{featbarplot}
\title{Feature extracted bar plots}
\usage{
featbarplot(
  data = "data",
  variable = "LA, mP, pP, TF",
  timepoints = "c() vector"
)
}
\arguments{
\item{data}{the results values as a long variable format as set up using results_pivot()}

\item{variable}{the intended variable to be examined}

\item{timepoints}{a vector containing the index values to be plotted on the x axis}
}
\value{

}
\description{
This function creates barplots to compare features at different time instances in the tekscan recording
}
\examples{
\dontrun{
featbarplot()
}
}
