% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncronise_functions.R
\name{find_peaks}
\alias{find_peaks}
\title{Peak finding helper function}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{values to be checked}

\item{m}{the range of values either side of each peak which must be different to identify a peak}
}
\value{

}
\description{
This function finds peaks in data and returns the numerical position in the vector
}
\examples{
\dontrun{
find_peaks()
}
}
