% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncronise_functions.R
\name{indexer}
\alias{indexer}
\title{Index matching function}
\usage{
indexer(
  sim = simextract(),
  tek = tekextract(),
  numcycle = 5,
  peaknum = 1,
  mod = 0
)
}
\arguments{
\item{sim}{Output of the sim extractor function. Defaults to calling simextract()}

\item{tek}{Output of the TekScan extractor function. Defaults to calling tekextract()}

\item{numcycle}{a variable telling the operation how many cycles were recorded}

\item{peaknum}{identifies which peak to use based on its position in the output vector of find_peaks}

\item{mod}{a custom correction variable for minor adjustment of the position.}
}
\value{

}
\description{
This function uses the find_peaks function to match tekscan output and tekscan frames to the simulator index
This function has two outputs and therefore cannot be used directly but must be called alone then the outputs
collected from the global environment. This is used automatically in the autofit program, but can be run manually.
}
\examples{
\dontrun{
indexer()
}
}
