% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{metric_labeller}
\alias{metric_labeller}
\title{Metric labels}
\usage{
metric_labeller(variable, value)
}
\arguments{
\item{variable}{variable from plot}

\item{value}{value from plot}
}
\value{

}
\description{
A sub function to organise the labels correctly. Not to be used alone
}
\examples{
\dontrun{
metric_labeller()
}
}
