% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{modplot}
\alias{modplot}
\title{Model fit plot}
\usage{
modplot(tek = tek_sum, sim = sim_sum, modselect = model)
}
\arguments{
\item{tek}{TekScan summary from tidytek() or autofit()}

\item{sim}{Simulator summary from simextract()}

\item{modselect}{model output of calmodel() or autofit()}
}
\value{

}
\description{
Graphically and interactively illustrates the fit of the calibration curve.
Useful graphical tool to check autofit() has correctly minimised R squared and fit the data
Best combined with syncplot() as two halves of the same analysis, using plotly.
Can be partial bundled.
}
\examples{
\dontrun{
modplot()
subplot(syncplot(), modplot(), nrows=2) \%>\% partial_bundle()
}
}
