% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{prof_peaks}
\alias{prof_peaks}
\title{Profile peaks locator}
\usage{
prof_peaks(df = data_join(), s = "Medial", plot = TRUE)
}
\arguments{
\item{df}{Data containing the simulator output}

\item{s}{a string to filter the table by for each condyle}

\item{plot}{a true/false switch which determines if the output of this function is graphical or numerical}
}
\value{

}
\description{
Similar to Syncplot, this function finds the peaks in the profile data and marks them for comparison
}
\examples{
\dontrun{
prof_peaks()
}
}
