% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{simext_full}
\alias{simext_full}
\title{Full Simulator profile extractor}
\usage{
simext_full(file = simfile, skiptorow = 3)
}
\arguments{
\item{file}{simulator file paths, defaults to the default expected global environment variable for simfiles}

\item{skiptorow}{as with simextract() this variable controls the headspace skipping until relevant cycles}
}
\value{

}
\description{
As distinct from simextract(), this version extracts the whole simulator output rather than a summary
}
\examples{
\dontrun{
simext_full()
}
}
