% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{simextract}
\alias{simextract}
\title{Simulator AF summary data extractor}
\usage{
simextract(file = simfile, skiptorow = 3)
}
\arguments{
\item{file}{a file path as a tab-delimited .txt.
Defaults to an empty variable for the string}

\item{skiptorow}{a number which indicates the row to begin reading from}
}
\value{

}
\description{
This function extracts the axial force data from a single station knee simulator file
}
\examples{
\dontrun{
simextract()
}
}
