% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{sumdataframe}
\alias{sumdataframe}
\title{Output Table}
\usage{
sumdataframe(file = sampleextract(), knee = "rtkn", calibrationmodel = model)
}
\arguments{
\item{file}{TekScan recording. Defaults to the output of sampleextract()}

\item{knee}{right or left knee, dictates labelling}

\item{calibrationmodel}{model for calibration as in calibrate()}
}
\value{

}
\description{
Summary of the calibrated data as a table. Suited for storage without rounding for future use
}
\examples{
\dontrun{
sumdataframe()
}
}
