% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{sumtab}
\alias{sumtab}
\title{Results table}
\usage{
sumtab(df = sumdataframe())
}
\arguments{
\item{df}{a dataframe to be tidied up for knit, defaults to sumdataframe() output.}
}
\value{

}
\description{
Summary of the calibrated data as a table. Suited for presentation with rounding and detailed headers
Includes slicers for interaction after knitting
Could be a wrapper function for sumdataframe?
}
\examples{
\dontrun{
sumtab(sumdataframe())
}
}
