% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_functions.R
\name{tekextract}
\alias{tekextract}
\title{Tekscan raw summary extractor}
\usage{
tekextract(file = tekfile, headerrows = 24)
}
\arguments{
\item{file}{a file path as a comma separated .csv.
Defaults to an empty variable for the string}

\item{headerrows}{rows to skip removing pre-amble. Defaults to 24 for an unsaved file}
}
\value{

}
\description{
This function extracts the raw data from a tekscan ascii file and creates an aggregate
}
\examples{
\dontrun{
tekextract()
}
}
