% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{tekplot2}
\alias{tekplot2}
\title{Animated Tekscan Plot}
\usage{
tekplot2(
  file = sampleextract(),
  calmodel = model,
  knee = "rtkn",
  index_labels = tek_ind
)
}
\arguments{
\item{file}{TekScan recording to be analysed. Defaults to the output of sampleextract()}

\item{calmodel}{Model for calibration of data. Defaults to the output of autofit()}

\item{knee}{right or left knee, as this dictates labelling}

\item{index_labels}{output from the Indexer function or autofit. Indicates gait progression on each frame.}
}
\value{

}
\description{
This is the big one. Takes the TekScan dynamic recording, calibrates it according to model.
Produces as an output raster plots of the recording including summary statistics, labels and
centre-of-force markers.
Outputs a complex plotly object with several animated graphs.
Requires a large amount of memory to Knit to HTML.
}
\examples{
\dontrun{
tekplot2()
}
}
