% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncronise_functions.R
\name{tidytek}
\alias{tidytek}
\title{Tidy tekscan data extractor}
\usage{
tidytek(x = indexer(), SumOrInd = "S")
}
\arguments{
\item{x}{Output of the indexer function, regardless of if the peaks match}

\item{SumOrInd}{string serving as a switch to produce either output}
}
\value{

}
\description{
Wrapper function using the indexer tool output to produce important summary data frames.
Has two possible outputs including a summary of the total value for each frame, or a each frame with its corresponding index values
}
\examples{
\dontrun{
tidytek()
}
}
