% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{tks_annotate}
\alias{tks_annotate}
\title{Tekscan peaks annotator}
\usage{
tks_annotate(
  P = tks_peaks(m1 = NULL),
  m2 = tks_peaks(plot = FALSE, m1 = NULL),
  res_long = rl
)
}
\arguments{
\item{P}{a plot generated by tks_peaks()}

\item{m2}{a list vector of peaks generated by tks_peaks(plot=NULL, m1=NULL)}

\item{res_long}{a dataframe of the results in a long format generated by results_pivot}
}
\value{

}
\description{
A further set of annotations generated from tks_peaks can be optionally annotated by generating the plot
from tks_peaks and passing it to this function.
}
\examples{
\dontrun{
tks_annotate()
}
}
