% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{tks_peaks}
\alias{tks_peaks}
\title{Tekscan plotter}
\usage{
tks_peaks(df = results_pivot(), plot = TRUE, m1 = NULL)
}
\arguments{
\item{df}{Data produced by results_pivot()}

\item{plot}{a true/false switch which determines if the output of this function is graphical or numerical}

\item{m1}{an optional list vector of peak locations generated by prof_peaks which annotates peak positions}
}
\value{

}
\description{
This function finds the peaks in the tekscan recording data and marks them for comparison
}
\examples{
\dontrun{
tks_peaks()
}
}
