% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_functions.R
\name{results_pivot}
\alias{results_pivot}
\title{Results long format}
\usage{
results_pivot(df = results)
}
\arguments{
\item{df}{results table or list of results tables to be produced.}
}
\value{

}
\description{
Takes the calibrated results, or optionally results list and reformats it to a long data structure to facilitate
plotting. Uses the "Position" variable as identifiers. Labels the variables in a standard format for other functions.
}
\examples{
\dontrun{
results_pivot()
}
}
